@@ -1,17 +1,17 @@
 name: flutter_clean_architecture
 description: A Flutter package that implements the Clean Architecture by Uncle Bob in Flutter. It provides Views, Controllers, Presenters, Observers, and UseCases.
-version: 5.0.3
+version: 6.0.0
 homepage: https://github.com/ShadyBoukhary/flutter_clean_architecture
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   logging: ^1.0.0
   meta: ^1.3.0
-  provider: ^6.0.1
+  provider: ^6.1.2
   responsive_builder: ^0.4.1
   rxdart: ^0.26.0
 
