@@ -5,9 +5,9 @@ import 'package:flutter_clean_architecture/flutter_clean_architecture.dart';
 
 class HomeController extends Controller {
   int _counter;
-  User _user;
+  User? _user;
   int get counter => _counter;
-  User get user => _user; // data used by the View
+  User? get user => _user; // data used by the View
   final HomePresenter homePresenter;
   // Presenter should always be initialized this way
   HomeController(usersRepo)
