@@ -2,7 +2,7 @@ import 'package:example/src/app/pages/home/home_controller.dart';
 import 'package:example/src/app/widgets/button.dart';
 import 'package:example/src/data/repositories/data_users_repository.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_clean_architecture/flutter_clean_architecture.dart';
+import 'package:flutter_clean_architecture/flutter_clean_architecture.dart' as clean;
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
@@ -25,16 +25,16 @@ void main() {
   });
 }
 
-class TestPage extends View {
+class TestPage extends clean.View {
   const TestPage({
-    Key key,
+    Key? key,
   }) : super(key: key);
 
   @override
   TestPageState createState() => TestPageState();
 }
 
-class TestPageState extends ViewState<TestPage, HomeController> {
+class TestPageState extends clean.ViewState<TestPage, HomeController> {
   TestPageState() : super(HomeController(DataUsersRepository()));
 
   @override
